requirements.txt
websockets>=10.0
aiohttp>=3.8
pyperclip>=1.8
PyYAML>=6.0
# macOS specific dependencies (only strictly needed on macOS)
pyobjc-core>=8.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=8.0; sys_platform == 'darwin'
python-socks>=0.2.1